Move agent tool creation code into appropriate directory
